+++ src/Platforms/Symbian/UtestPlatform.cpp
@@ -34,8 +34,8 @@
 
 void Utest::executePlatformSpecificTestBody()
 {
-		TInt err(KErrNone); 
-		TRAP(err, testBody()); 
+		TInt err(KErrNone);
+		TRAP(err, testBody());
 		if(err != KErrNone) {
 			Utest::getCurrent()->fail("Leave in test method", "", 0);
 		}
@@ -48,7 +48,7 @@ static long TimeInMillisImplementation()
 	struct timeval tv;
 	struct timezone tz;
 	::gettimeofday(&tv, &tz);
-	return (tv.tv_sec * 1000) + (long)(tv.tv_usec * 0.001);	
+	return (tv.tv_sec * 1000) + (long)(tv.tv_usec * 0.001);
 }
 
 static long (*timeInMillisFp) () = TimeInMillisImplementation;
@@ -94,18 +94,18 @@ void TestRegistry::platformSpecificRunOneTest(Utest* test, TestResult& result)
         runOneTest(test, result) ;
 }
 
-void PlatformSpecificExitCurrentTestImpl() 
+void PlatformSpecificExitCurrentTestImpl()
 {
     longjmp(test_exit_jmp_buf, 1);
 }
 
-void FakePlatformSpecificExitCurrentTest() 
+void FakePlatformSpecificExitCurrentTest()
 {
 }
 
 void (*PlatformSpecificExitCurrentTest)() = PlatformSpecificExitCurrentTestImpl;
 
-int PlatformSpecificSprintf(char *str, size_t size, const char *format, ...) 
+int PlatformSpecificSprintf(char *str, unsigned int size, const char *format, ...)
 {
    va_list args;
    va_start(args, format);
